requirements.txt
python-dotenv~=1.0.0
jira~=3.5.0
prettytable~=3.7.0
